requirements.txt
PyPDF2==3.0.1
python-dateutil==2.8.2
Pillow>=10.0.0
google-cloud-vision==3.4.4
pdf2image==1.16.3
opencv-python>=4.8.0
